modify test bug

+++ include/linux/syscalls.h
@@ -902,6 +902,6 @@ asmlinkage long sys_pkey_mprotect(unsigned long start, size_t len,
 				  unsigned long prot, int pkey);
 asmlinkage long sys_pkey_alloc(unsigned long flags, unsigned long init_val);
 asmlinkage long sys_pkey_free(int pkey);
-asmlinkage int sys_test(char *str, int count);
+asmlinkage long sys_test(char *str, int count);
 
 #endif
